Refactor logic for deriving GitHub issue tags
Organize label-related logic into a single Node.js module. For the
purposes of code clarity, change references to "raw labels" to instead
describe "file labels." Remove the `rawLabels` metadata property,
persisting only the complete `labels` property. Refactor internal
functions to accept only the data necessary to perform their task. Add
documentation.
