+++ lib/metadata/index.js
@@ -2,7 +2,6 @@
 
 var filenames = require('./filenames'),
     paths = require('./paths'),
-    rawLabels = require('./raw-labels'),
     findSpecs = require('./find-specs'),
     findOwners = require('./find-owners'),
     findRemovedReviewers = require('./find-removed-reviewers'),
@@ -20,12 +19,14 @@ module.exports = function getMetadada(number, author, content) {
     };
     author = author.toLowerCase();
     var reviewers;
+    var fileLabels;
+
     return filenames(number)
         .then(function(filenames) {
             metadata.filenames = filenames.all;
             metadata.filenamesIgnoreRemoved = filenames.ignoreRemoved;
             metadata.paths = paths(metadata.filenames);
-            metadata.rawLabels = rawLabels(metadata.filenames);
+            fileLabels = labels.fromFiles(metadata.filenames);
             metadata.isRoot = metadata.filenames.some(function(path) {
                 return path.split('/').length == 1
             });
@@ -33,11 +34,13 @@ module.exports = function getMetadada(number, author, content) {
                 var dir = path.split('/')[0];
                 return dir == "tools" || dir == "resources";
             });
-            return findSpecs(metadata.rawLabels);
+            return findSpecs(fileLabels);
         }).then(function(specs) {
             metadata.specs = specs;
             metadata.workingGroups = wg(specs);
-            metadata.labels = labels(metadata);
+            metadata.labels = fileLabels.concat(
+                labels.fromWorkingGroups(metadata.workingGroups)
+            );
         }).then(function() {
             return findOwners(metadata.paths)
         }).then(function(owners) {
